add a jupyter-notebook Terra plugin
This implements jupyter-notebook as a workload template.

It was tested on the default docker stacks from Jupyter - that means we
use the JupyterLab frontend and rely on its config options (https://jupyter-docker-stacks.readthedocs.io/en/latest/index.html)

This was tested both on the upcoming traefik-based httproutes and a
stable 1.4.0 ingress-based deployment.

This was tested with both persistant and non-persistant home
directories (non-persistant is still useful for testing/POC/demos)